0618 1111

+++ testModel01/testModel01/BackEnd/BS_床位編輯.aspx.cs
@@ -14,7 +14,7 @@ namespace testModel01.BackEnd
 
         string str_Data_Source = WebConfigurationManager.OpenWebConfiguration("/testModel01").ConnectionStrings.ConnectionStrings["dbKSYYConnectionString"].ConnectionString;//家2
 
-        DropDownList ddlist;  //暫存
+        DropDownList[] ddlist = new DropDownList[3];  //暫存
         Image img1;//暫存
         FileUpload fupload; //暫存
         Control contrl_temp;
@@ -33,28 +33,31 @@ namespace testModel01.BackEnd
             string str_bednum= Request.QueryString["f床號"].ToString();
           
             SqlDataSource sds = new SqlDataSource();
-            //sds.ConnectionString = "Data Source=" + str_Data_Source + ";Initial Catalog=dbKSYY;Integrated Security=True";
             sds.ConnectionString = str_Data_Source;
             sds.SelectCommand = "SELECT * from  T床位系統 WHERE f床號 ='"+str_bednum+"'";
             DataView dn = (DataView)sds.Select(DataSourceSelectArguments.Empty);            
             int[] inta_ = new int[3];
 
-         try
-         {
-             m_FindControl_ddlist(FormView16, typeof(DropDownList));
-             ddlist.SelectedIndex = Convert.ToInt32(dn[0]["f是否住院中"]);
-         }
-         catch (Exception exc)
-         {
-             ddlist = new DropDownList();
-             ddlist.SelectedIndex = 0;
-         }
+            //dropdownlist的全部選擇
+            m_FindControl_ddlist(FormView16, typeof(DropDownList));
+         
+                ddlist[0].SelectedIndex = Convert.ToInt32(dn[0]["f是否住院中"]);
+                ddlist[1].SelectedIndex = Convert.ToInt32(dn[0]["f鼻胃管"]);
+                ddlist[2].SelectedIndex = Convert.ToInt32(dn[0]["f導尿管"]);
+                 try
+            {
+            }
+            catch (Exception exc)
+            {
+                ddlist[0] = ddlist[1] = ddlist[2] = new DropDownList();
+                ddlist[0].SelectedIndex = 0;
+            }
 
-
+    
          try
          {
              img1 = new Image();  
-             img1.ImageUrl = dn[0]["f大頭照"].ToString();
+             img1.ImageUrl = dn[0]["f照片"].ToString();
              if (img1.ImageUrl.Length < 5)
              {
                
@@ -77,9 +80,17 @@ namespace testModel01.BackEnd
             {
                 if (((DropDownList)root).ID == "DDList_hospital")
                 {
-                    ddlist = ((DropDownList)root);
-                    return;
+                    ddlist[0] = ((DropDownList)root);                    
+                }else
+                if (((DropDownList)root).ID == "DDList_nose")
+                {
+                    ddlist[1] = ((DropDownList)root);
                 }
+                else
+                    if (((DropDownList)root).ID == "DDList_pee")
+                    {
+                        ddlist[2] = ((DropDownList)root);
+                    }
             }
             foreach (Control c in root.Controls)
                 m_FindControl_ddlist(c, type);
@@ -112,8 +123,21 @@ namespace testModel01.BackEnd
         protected void FormView16_ItemUpdating(object sender, FormViewUpdateEventArgs e)
         {
             FormView f = (FormView)sender;
+
+
             m_FindControl_ddlist(FormView16, typeof(DropDownList));
-            e.NewValues["f是否住院中"] = ddlist.SelectedIndex.ToString();
+            try
+            {
+
+                e.NewValues["f是否住院中"] = ddlist[0].SelectedIndex.ToString();
+                e.NewValues["f鼻胃管"] = ddlist[1].SelectedIndex.ToString();
+                e.NewValues["f導尿管"] = ddlist[2].SelectedIndex.ToString();
+
+            }
+            catch (Exception exc)
+            {
+         
+            }
 
             
             //照片
@@ -146,11 +170,11 @@ namespace testModel01.BackEnd
                     }
                     //上傳照片                   
                         fupload.SaveAs(Server.MapPath(@"~\pic\床位照片\大頭照\" + fupload.FileName));
-                        e.NewValues["f大頭照"] = @"..\pic\床位照片\大頭照\" + fupload.FileName;
+                        e.NewValues["f照片"] = @"..\pic\床位照片\大頭照\" + fupload.FileName;
                  }
             }
            else
-               e.NewValues["f大頭照"] = img1.ImageUrl;
+               e.NewValues["f照片"] = img1.ImageUrl;
         }
 
         protected void FormView16_ItemDeleted(object sender, FormViewDeletedEventArgs e)
@@ -168,11 +192,13 @@ namespace testModel01.BackEnd
             m_FindControl_Control(FormView16, typeof(Image));
             img1 = ((Image)contrl_temp);
 
-            if (img1.ImageUrl.Substring(img1.ImageUrl.LastIndexOf(@"/\") + 1).ToLower()
+            if (img1.ImageUrl.Substring(img1.ImageUrl.LastIndexOf(@"/") + 1).ToLower()
+                     != "defaultimg.png" &&
+                img1.ImageUrl.Substring(img1.ImageUrl.LastIndexOf(@"\") + 1).ToLower()
                      != "defaultimg.png")
             {
                 string str_expath = System.Web.Hosting.HostingEnvironment.MapPath("~");
-                string str_temp = str_expath + img1.ImageUrl.Substring(4, img1.ImageUrl.Length - 4);
+                string str_temp = str_expath + img1.ImageUrl.Substring(3, img1.ImageUrl.Length - 3);
                 if (System.IO.File.Exists(str_temp))
                     System.IO.File.Delete(str_temp);
             }
@@ -182,6 +208,35 @@ namespace testModel01.BackEnd
         protected void Button1_Click(object sender, EventArgs e)
         {
             Response.Redirect(this.Page.Request.UrlReferrer.ToString());
+        }
+
+        protected void Button2_Click(object sender, EventArgs e)
+        {
+          
+        }
+
+        protected void btn_delpic_Click1(object sender, EventArgs e)
+        {
+            m_FindControl_Control(FormView16, typeof(Image));
+            img1 = ((Image)contrl_temp);
+
+            if (img1.ImageUrl.Substring(img1.ImageUrl.LastIndexOf(@"/") + 1).ToLower()
+                     != "defaultimg.png" &&
+                img1.ImageUrl.Substring(img1.ImageUrl.LastIndexOf(@"\") + 1).ToLower()
+                     != "defaultimg.png")
+            {
+                string str_expath = System.Web.Hosting.HostingEnvironment.MapPath("~");
+                string str_temp = str_expath + img1.ImageUrl.Substring(3, img1.ImageUrl.Length - 3);
+                if (System.IO.File.Exists(str_temp))
+                    System.IO.File.Delete(str_temp);
+       
+
+            SqlDataSource sds = new SqlDataSource();
+            sds.ConnectionString = str_Data_Source;
+            sds.UpdateCommand = "update T床位系統 set f照片 = null WHERE f照片 ='" + img1.ImageUrl + "'";
+            sds.Update();
+            Response.Redirect(this.Page.Request.UrlReferrer.ToString());
+            }
         }
